// src/types/websiteListing.ts
export class websiteListing {
  orgCode: string | null;
  orgName: string | null;
  websiteName: string | null;
  websiteNameStripped: string | null;
  displayName: string | null;
  guid: string | null;
  websiteID: number | null;
  isCurrent: boolean | null;
  managerUrl: string | null;
  version: string | null;
  isOwner: boolean | null;
  isDormant: boolean | null;
  isRestoring: boolean | null;
}